installer: split plan.rs into action and plan
